Pull fluentd event time to its own header
