+++ client_out.txt
@@ -1,0 +1,1 @@
+python3: can't open file '/home/cjen1-msft/CCF/.worktrees/msgpack/probe_client.py': [Errno 2] No such file or directory
